requirements.txt
langchain
googlesearch.py
python-dotenv
requests